% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidProfillingRMpF.R
\name{ks_pvalues}
\alias{ks_pvalues}
\title{return ks_test p value between two numeric groups in data frame}
\usage{
ks_pvalues(Th, DF, colNameControl, conGroup, altGroup, colNameNumeric,
  Above = TRUE)
}
\arguments{
\item{Th}{(Value)}

\item{DF}{(data frame)}

\item{colNameControl}{(string - col name containing group information)}

\item{conGroup}{(string - control group)}

\item{altGroup}{(string - alternative group)}

\item{colNameNumeric}{(string - col name of numeric variable to compare)}

\item{Above}{(bool - use values above or below)}
}
\value{
double
}
\description{
return ks_test p value between two numeric groups in data frame
}
\examples{
---
}
