% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidProfillingRMpF.R
\name{plotKSpvalues}
\alias{plotKSpvalues}
\title{plot KS_test p values}
\usage{
plotKSpvalues(Range, DF, colNameControl, conGroup, altGroup,
  colNameNumeric, Above, ...)
}
\arguments{
\item{Range}{(vector - e.g. 25:70)}

\item{DF}{(data frame)}

\item{colNameControl}{(string - col name containing group information)}

\item{conGroup}{(string - control group)}

\item{altGroup}{(string - alternative group)}

\item{colNameNumeric}{(string - col name of numeric variable to compare)}

\item{Above}{(bool - use values above or below)}
}
\value{
double
}
\description{
plot KS_test p values
}
\examples{
---
}
