% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidProfillingRMpF.R
\name{RocWith_ci}
\alias{RocWith_ci}
\title{Plot ROC curve with 95% ci}
\usage{
RocWith_ci(obj, legacy = TRUE)
}
\arguments{
\item{obj}{(object from ROC package)}

\item{legacy}{(bool - plot order on X axis - legacy equal true correspond to same plot order as legacy software)}
}
\description{
Plot ROC curve with 95% ci
}
\examples{
---
}
